Add text hover animations for menu buttons
Added a hover animation that grow a font by a percent controled by an
animation curve. When that curve is done and the mouse remains hovering
there is a pulse animation curve that occurs until the mouse leaves.
Added an editor drawer for the ui script to modify event trigger class
for the animation.